Moved loader to cpp
